# ProyectoFinalBackend/src/services/swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"version": "1.0.0",
		"title": "Backend Project",
		"description": "<b>CoderHouse Backend Project</b>.",
		"contact": { "name": "Facundo Duhalde", "email": "facundo_duhalde@hotmail.com" }
	},
	"servers": [
		{ "url": "http://localhost:8080", "description": "Development server" }
	],
	"basePath": "/",
	"schemes": ["http", "https"],
	"paths": {
...
		"/api/image/upload": {
			"post": {
				"tags": ["Image"],
				"security": [{ "bearerAuth": [] }],
				"description": "Subir imagen de producto. Utilizar Postman, seleccionando el formato JPG/PNG.",
				"requestBody": {
					"content": {
						"image/png:": {
							"schema": {
								"type": "string",
								"format": "binary"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Success",
						"example": "http://localhost:8080/api/image/61172bf8387c613019fe6b85"
					},
					"401": { "description": "Unauthorized" }
				}
			}
		},
		"/api/image/{id}": {
			"get": {
				"tags": ["Image"],
				"description": "Buscar imagen por su id",
				"parameters": [
					{ "name": "id", "in": "path", "required": true, "type": "string" }
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"image/png": {
								"schema": {
									"type": "string",
									"format": "binary"
								}
							}
						}
					},
					"404": {
						"description": "Not found"
					}
				}
			}
		},
		"/api/image/{filename}": {
			"delete": {
				"tags": ["Image"],
				"security": [{ "bearerAuth": [] }],
				"description": "Borrar imagen, informar nombre completo",
				"parameters": [
					{
						"name": "filename",
						"in": "path",
						"required": true,
						"type": "string",
						"example": "http://localhost:8080/api/image/poroto.jpeg"
					}
				],
				"responses": {
					"200": { "description": "Se elimino la imagen" },
					"401": { "description": "Unauthorized" }
				}
			}
		},
...
	},
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		},
...
	}
}